Condensed most tests testing same method into a single one Made a couple changes here n there

+++ SportCenter-Backend/src/main/java/ca/mcgill/ecse321/Sport/Center/Application/ECSE321/service/SchedulingService.java
@@ -101,7 +101,11 @@ public class SchedulingService {
     }
     @Transactional
     public void deleteSession(int sessionId){
-        sessionRegistrationRepository.deleteAllBySessionId(sessionId);
+        if(!sessionRepository.existsById(sessionId)){
+            throw new IllegalArgumentException("No session with given ID");
+        }
+        
+        sessionRegistrationRepository.deleteBySessionId(sessionId);
         sessionRepository.deleteById(sessionId);
     }
